[Heatmap] Adjusting grid calcualtion for consistent grid layout
Date math is hard!

This addresses a problem in how the heatmap lays calculates days to lay out the grid. The calculations were being thrown off in some scenarios by a combination of factors including timezone offsets, time of day, and daylight savings time.

This patch should correct all this by forcing each day to be midnight UTC when laying out the grid.